refactor: move groups components into features with stubs
