+++ src/features/groups/pages/GroupDetailPage.tsx
@@ -1,4 +1,4 @@
 // GroupDetailPage - Re-exporta el Container para mantener compatibilidad de imports
 // Este archivo sirve como punto de entrada para el patrón Container/Presentational
 
-export { GroupDetailContainer as default } from '@/components/GroupDetailContainer';
+export { GroupDetailContainer as default } from '@/features/groups/components/GroupDetailContainer';
